Api router refactoring (#121)
* Router refactoring: Application

* Minor: rename test for test_router_application

* Refactor router for user and test

* Refactor for router role
